CHAD-17084: zwave-bulb lazy loading of sub-drivers
