+++ drivers/SmartThings/zwave-bulb/src/aeon-multiwhite-bulb/fingerprints.lua
@@ -1,0 +1,10 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local AEON_MULTIWHITE_BULB_FINGERPRINTS = {
+  {mfr = 0x0371, prod = 0x0103, model = 0x0001}, -- Aeon LED Bulb 6 Multi-White US
+  {mfr = 0x0371, prod = 0x0003, model = 0x0001}, -- Aeon LED Bulb 6 Multi-White EU
+  {mfr = 0x0300, prod = 0x0003, model = 0x0004}  -- ilumin Tunable White
+}
+
+return AEON_MULTIWHITE_BULB_FINGERPRINTS
